Add the dish details in the DetailedView (formatting pending)
